Add mxu1_as_macros.s.h header
This is the central header, all that is needed for inclusion into a
project that wishes to assemble MXU1 rev1 or rev2 opcodes.

For now, there is no ability to restrict whether it is just rev1 opcodes
that are allowed to be generated with the header: The MXU documentation
does state which ones were added for rev2, but it is unknown how accurate
these docs are.

Signed-off-by: Daniel Silsby <dansilsby@gmail.com>